[chore] cleanup tests and move examples to own file
