+++ config/confighttp/compression_test.go
@@ -8,12 +8,14 @@ import (
 	"compress/gzip"
 	"compress/zlib"
 	"context"
+	"errors"
 	"fmt"
 	"io"
 	"net/http"
 	"net/http/httptest"
 	"strings"
 	"testing"
+	"testing/iotest"
 
 	"github.com/golang/snappy"
 	"github.com/klauspost/compress/zstd"
@@ -142,7 +144,7 @@ func TestHTTPCustomDecompression(t *testing.T) {
 	require.NoError(t, err, "failed to create request to test handler")
 	req.Header.Set("Content-Encoding", "custom-encoding")
 
-	client := http.Client{}
+	client := srv.Client()
 	res, err := client.Do(req)
 	require.NoError(t, err)
 
@@ -260,7 +262,7 @@ func TestHTTPContentDecompressionHandler(t *testing.T) {
 			require.NoError(t, err, "failed to create request to test handler")
 			req.Header.Set("Content-Encoding", tt.encoding)
 
-			client := http.Client{}
+			client := srv.Client()
 			res, err := client.Do(req)
 			require.NoError(t, err)
 
@@ -287,7 +289,7 @@ func TestHTTPContentCompressionRequestWithNilBody(t *testing.T) {
 	req, err := http.NewRequest(http.MethodGet, server.URL, nil)
 	require.NoError(t, err, "failed to create request to test handler")
 
-	client := http.Client{}
+	client := server.Client()
 	client.Transport, err = newCompressRoundTripper(http.DefaultTransport, configcompression.TypeGzip)
 	require.NoError(t, err)
 	res, err := client.Do(req)
@@ -298,27 +300,16 @@ func TestHTTPContentCompressionRequestWithNilBody(t *testing.T) {
 	require.NoError(t, res.Body.Close(), "failed to close request body: %v", err)
 }
 
-type copyFailBody struct {
-}
-
-func (*copyFailBody) Read([]byte) (n int, err error) {
-	return 0, fmt.Errorf("read failed")
-}
-
-func (*copyFailBody) Close() error {
-	return nil
-}
-
 func TestHTTPContentCompressionCopyError(t *testing.T) {
 	server := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, _ *http.Request) {
 		w.WriteHeader(http.StatusOK)
 	}))
 	t.Cleanup(server.Close)
 
-	req, err := http.NewRequest(http.MethodGet, server.URL, &copyFailBody{})
+	req, err := http.NewRequest(http.MethodGet, server.URL, iotest.ErrReader(errors.New("read failed")))
 	require.NoError(t, err)
 
-	client := http.Client{}
+	client := server.Client()
 	client.Transport, err = newCompressRoundTripper(http.DefaultTransport, configcompression.TypeGzip)
 	require.NoError(t, err)
 	_, err = client.Do(req)
@@ -342,7 +333,7 @@ func TestHTTPContentCompressionRequestBodyCloseError(t *testing.T) {
 	req, err := http.NewRequest(http.MethodGet, server.URL, &closeFailBody{Buffer: bytes.NewBuffer([]byte("blank"))})
 	require.NoError(t, err)
 
-	client := http.Client{}
+	client := server.Client()
 	client.Transport, err = newCompressRoundTripper(http.DefaultTransport, configcompression.TypeGzip)
 	require.NoError(t, err)
 	_, err = client.Do(req)
@@ -362,7 +353,7 @@ func TestOverrideCompressionList(t *testing.T) {
 	require.NoError(t, err, "failed to create request to test handler")
 	req.Header.Set("Content-Encoding", "snappy")
 
-	client := http.Client{}
+	client := srv.Client()
 
 	// test
 	res, err := client.Do(req)
