+++ config/confighttp/confighttp_test.go
@@ -1011,28 +1011,6 @@ func verifyHeadersResp(t *testing.T, url string, expected map[string]configopaqu
 	}
 }
 
-func ExampleServerConfig() {
-	settings := NewDefaultServerConfig()
-	settings.Endpoint = "localhost:443"
-
-	s, err := settings.ToServer(
-		context.Background(),
-		componenttest.NewNopHost(),
-		componenttest.NewNopTelemetrySettings(),
-		http.HandlerFunc(func(http.ResponseWriter, *http.Request) {}))
-	if err != nil {
-		panic(err)
-	}
-
-	l, err := settings.ToListener(context.Background())
-	if err != nil {
-		panic(err)
-	}
-	if err = s.Serve(l); err != nil {
-		panic(err)
-	}
-}
-
 func TestHttpClientHeaders(t *testing.T) {
 	tests := []struct {
 		name    string
